+++ R/ForestTools.R
@@ -2,8 +2,8 @@
 #'
 #' Forest Tools provides functions for analyzing remotely sensed forest data.
 #' Functions like \link{TreeTopFinder} and \link{SegmentCrowns} are applied to rasterized
-#' canopy height models, and can detect treetops and outline their respective crowns. \link{TreeTopSummary} can
-#' summarize tree counts and tree attributes within particular areas of interest or within continous spatial
+#' canopy height models, and can detect treetops and outline their respective crowns. \link{SpatialStatistics} can
+#' summarize tree counts and attributes within particular areas of interest or within continous spatial
 #' grids.
 #'
 #' @docType package