+++ NEWS.md
@@ -1,3 +1,9 @@
+# ForestTools 0.1.1
+
+Modified 'SegmentCrowns' function so that it can produce tree crowns in polygon format. The function will also calculate crown area, and filter out crowns that are not associated with a treetop point location.
+
+Changed name of 'TreeTopSummary' function to 'SpatialStatistics'. This reflects its new functionality, which allows crown maps to be inputted, as well as treetop locations. 
+
 # ForestTools 0.1.0
 
 * Initial release of ForestTools